requirements.txt
pydantic
pyyaml
requests
instructor
openai
python-dotenv
requests